+++ src/vuer/__init__.py
@@ -2,11 +2,10 @@ from vuer._compat import handle_server_import_error
 
 try:
     from vuer.server import Vuer, VuerSession
-    from vuer.workspace import Blob, Workspace, jpg, png, b64jpg, b64png
+    from vuer.workspace import Blob, Workspace
 except ImportError as e:
     handle_server_import_error(e)
     Vuer, VuerSession, Workspace, Blob = None, None, None, None
-    jpg, png, b64jpg, b64png = None, None, None, None
 
 from vuer.client import VuerClient
 
@@ -23,9 +22,5 @@ __all__ = [
     "VuerClient",
     "Workspace",
     "Blob",
-    "jpg",
-    "png",
-    "b64jpg",
-    "b64png",
     "entrypoint",
 ]